♻️ refactor(server): collapse UserStatsUpdater to its lazy window-decay self-heal only
